Update main.py
The data are randomly shuffled to separate the training and test set, according to which the network will be trained and tested, respectively. The percentage of the training set corresponds to 80% of the data, while that of the test set, to the remaining 20% of the total data. In the pre-processing stage, the images are cropped to dimensions 224x224, categorized according to the class to which they belong and subjected to accidental alteration of some features, such as shift, inversion, focus, etc and arranged in a tabular form.